backend/prediction/requirements.txt
fastapi
uvicorn
requests
pandas
schedule
pyyaml
openpyxl
sklearn